+++ app/sprinkles/core/src/Core.php
@@ -10,7 +10,7 @@
 
 namespace UserFrosting\Sprinkle\Core;
 
-use Interop\Container\ContainerInterface;
+use Psr\Container\ContainerInterface;
 use RocketTheme\Toolbox\Event\Event;
 use UserFrosting\Sprinkle\Core\Csrf\SlimCsrfProvider;
 use UserFrosting\Sprinkle\Core\Database\Models\Model;