+++ app/sprinkles/core/src/Database/Seeder/Seeder.php
@@ -11,7 +11,7 @@
 namespace UserFrosting\Sprinkle\Core\Database\Seeder;
 
 use Illuminate\Support\Str;
-use Interop\Container\ContainerInterface;
+use Psr\Container\ContainerInterface;
 use UserFrosting\UniformResourceLocator\Resource as ResourceInstance;
 
 /**